Used uintptr_t for base address of hardware ports

+++ solutions/03_hardware_manipulation/src/main.cpp
@@ -6,7 +6,7 @@
 #include "Timer.h"       // sleep
 #include <cstdint>
 
-constexpr uint32_t GPIOD_base { 0x4002'0C00u };
+constexpr uintptr_t GPIOD_base { 0x4002'0C00u };
 
 static volatile uint32_t *const GPIOD_moder = reinterpret_cast<uint32_t*>(GPIOD_base);
 static volatile uint32_t *const GPIOD_odr = reinterpret_cast<uint32_t*>(GPIOD_base + 0x14u);
