+++ solutions/06_structuring_code/src/GPIO_C.h
@@ -10,7 +10,7 @@
 
 namespace Devices {
 
-constexpr uint32_t GPIOD_base { 0x4002'0C00u };
+constexpr uintptr_t GPIOD_base { 0x4002'0C00u };
 
 static volatile uint32_t *const GPIOD_moder = reinterpret_cast<uint32_t*>(GPIOD_base);
 static volatile uint32_t *const GPIOD_odr = reinterpret_cast<uint32_t*>(GPIOD_base + 0x14u);
